feat(refactor-widgets): refactored the widgets imports
