+++ src/components/Widgets/index.jsx
@@ -1,0 +1,26 @@
+/*
+ Copyright (C) 2021 Shruti Agarwal (mail2shruti.ag@gmail.com)
+
+ SPDX-License-Identifier: GPL-2.0
+
+ This program is free software; you can redistribute it and/or
+ modify it under the terms of the GNU General Public License
+ version 2 as published by the Free Software Foundation.
+ This program is distributed in the hope that it will be useful,
+ but WITHOUT ANY WARRANTY; without even the implied warranty of
+ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ GNU General Public License for more details.
+
+ You should have received a copy of the GNU General Public License along
+ with this program; if not, write to the Free Software Foundation, Inc.,
+ 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+*/
+
+import { Spinner } from "react-bootstrap";
+import Alert from "./Alert";
+import Button from "./Button";
+import Image from "./Image";
+import InputContainer from "./Input";
+import Tooltip from "./Tooltip";
+
+export { Alert, Button, Image, InputContainer, Tooltip, Spinner };
