+++ src/pages/Organize/Folder/Move/index.jsx
@@ -17,9 +17,7 @@
 */
 
 import React, { useState, useEffect } from "react";
-import InputContainer from "../../../../components/Widgets/Input";
-import Alert from "../../../../components/Widgets/Alert";
-import Button from "../../../../components/Widgets/Button";
+import { Alert, Button, InputContainer } from "../../../../components/Widgets";
 import {
   getAllFolders,
   moveFolder,
