+++ src/pages/Upload/File/index.jsx
@@ -17,10 +17,12 @@
 */
 
 import React, { useState, useEffect } from "react";
-import { Spinner } from "react-bootstrap";
-import InputContainer from "../../../components/Widgets/Input";
-import Alert from "../../../components/Widgets/Alert";
-import Button from "../../../components/Widgets/Button";
+import {
+  Alert,
+  Button,
+  InputContainer,
+  Spinner,
+} from "../../../components/Widgets";
 import CommonFields from "../../../components/Upload/CommonFields";
 import { createUploadFile, scheduleJobs } from "../../../services/upload";
 import { getAllFolders } from "../../../services/folders";
